+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### Version 1.0.3
+- **Feature**: Add support of custom pagebuilder elements for categories [#4]
+- **Feature**: Add `CmsRowServiceLinks` element to pagebuilder to enable creation of service links [#3]
+- **Feature**: Add `CmsRowText` component to pagebuilder to enable creation of rich text AST elements [#2]
+
 ### Version 1.0.2
 - **Enhancement**: Remove dependency for `softcommerce/module-profile` in favour of `softcommerce/module-core`.